refactor: Move Reanimated components into root component folder
